2024.10.30 dependencies 추가, 공통으로 사용하는 SDK 추가하고, 네트워크 관련 공통 코드를 추가함. extensions 추가 및 코드 수정
