+++ Sources/Extensions/Array+Extension.swift
@@ -62,5 +62,9 @@ public extension Array {
         }
         return nil
     }
+
+    subscript(safe index: Int) -> Element? {
+        return indices.contains(index) ? self[index] : nil
+    }
 }
