+++ Sources/Extensions/String+Extension.swift
@@ -18,7 +18,6 @@ public extension String {
         // Index 값 획득
         let startIndex = index(self.startIndex, offsetBy: from)
         let endIndex = index(self.startIndex, offsetBy: to + 1) // '+1'이 있는 이유: endIndex는 문자열의 마지막 그 다음을 가리키기 때문
-        
         // 파싱
         return String(self[startIndex ..< endIndex])
     }
